Requisito 8 - Refatorei umas partes do requisito 7 para o codigo ficar um pouco mais legivel
