+++ front-end/src/components/componentsCSS/TopMenu.css
@@ -19,7 +19,7 @@
   width: 8%;
   height: 40px;
   background-color: #5fd450;
-  background-image: url(../icons/hamburguer-icon.png);
+  background-image: url(../../icons/hamburguer-icon.png);
   background-repeat: no-repeat;
   background-size: 60%;
   background-position: center;
